+++ tosixinch/tests/_test_actualrun.py
@@ -2,7 +2,7 @@
 
 """Actualy invoke tosixinch.main._main().
 
-For each rsrc from urls.txt,
+For each rsrc from rsrcs.txt,
 do download, extract, toc or convert.
 And compare the outputs with prepared reference files.
 
@@ -25,7 +25,7 @@ in application directory.
     .../tosixinch/tosixinch/tests/temp/actualrun/outcome
 
 '--creater-ref' changes current directory to 'reference',
-cleans files, prepares (copying urls.txt), and runs main._main().
+cleans files, prepares (copying rsrcs.txt), and runs main._main().
 The result is that it is populated with reference html and pdf files.
 
 Actual tests are always done with 'outcome' as current directory.
@@ -640,7 +640,7 @@ def parse_args(args=sys.argv[1:]):
         action='store_const', const='yes',
         help='print numbers and rsrcs')
     parser.add_argument('-n', '--number',
-        help='file number to process in urls.txt.')
+        help='file number to process in rsrcs.txt.')
 
     parser.add_argument('--create-ref',
         action='store_const', const='yes',